requirements.txt
# Use conda install ... to install these
matplotlib == 2.0.2
Pillow == 4.1.1
graphviz == 0.8 # for graphviz, installation via the system might be required, e.g. sudo apt-get install graphviz
pytest >= 3.0.7

# Use pip install ... to install these
keras == 2.0.6
tensorflow == 1.2.1
# tensorflow-gpu == 1.2.1 # Alternatively if available to run vastly faster trainings. See https://github.com/philferriere/dlwin
codecov == 2.0.9
pytest-cov == 2.5.1
h5py == 2.7.0
scikit-learn == 0.18.1
scipy == 0.19.0
sympy == 1.0
PyHamcrest == 1.9.0
muscima == 0.3.2
numpy == 1.12.1
lxml == 3.7.3
pydot == 1.2.3
tqdm
opencv-python
omrdatasettools==0.8

# Optional for reporting
telegram_send == 0.10
google-api-python-client == 1.6.2 # if pip install does not work, run "pip install --upgrade google-api-python-client"